build sh

+++ GViewCore/src/View/BufferViewer/Instance.cpp
@@ -333,7 +333,7 @@ void Instance::MoveToZone(bool startOfZone, bool select)
 bool Instance::ShowGoToDialog()
 {
     GoToDialog dlg(settings.get(), this->Cursor.currentPos, this->obj->GetData().GetSize());
-    if (dlg.Show() == (int) Dialogs::Result::Ok)
+    if (dlg.Show() == Dialogs::Result::Ok)
     {
         MoveTo(dlg.GetResultedPos(), false);
     }
